requirements.txt
alembic==1.12.0
annotated-types==0.5.0
anyio==3.7.1
asyncclick==8.1.3.4
asyncpg==0.28.0
bcrypt==4.0.1
boto3==1.28.44
botocore==1.31.44
click==8.1.7
databases==0.8.0
fastapi==0.103.1
flake8==6.1.0
greenlet==2.0.2
h11==0.14.0
httptools==0.6.0
idna==3.4
isort==5.12.0
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
mccabe==0.7.0
passlib==1.7.4
psycopg2-binary==2.9.7
pycodestyle==2.11.0
pydantic==2.3.0
pydantic_core==2.6.3
pyflakes==3.1.0
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
PyYAML==6.0.1
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.49
starlette==0.27.0
typing_extensions==4.7.1
urllib3==1.26.16
uvicorn==0.23.2
uvloop==0.17.0
watchfiles==0.20.0
websockets==11.0.3
